requirements.txt
# XAIDroid Requirements




# Android Analysis
androguard>=3.4.0
pyaxmlparser>=0.3.0

# Graph Processing
networkx>=3.0
# pygraphviz>=1.11  # For visualization (requires graphviz system package)

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Genetic Algorithm (GAM)
deap>=1.4.0  # Distributed Evolutionary Algorithms in Python

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Logging & Monitoring
tensorboard>=2.13.0
tqdm>=4.65.0
colorlog>=6.7.0

# Code Quality
black>=23.3.0
flake8>=6.0.0
pytest>=7.3.0
pytest-cov>=4.1.0

# Optional: Experiment Tracking
# wandb>=0.15.0

# Utilities
joblib>=1.3.0
dill>=0.3.6  # Better pickling for complex objects